rm unused action
