--- ts/features/pn/store/actions/service.ts
@@ -1,7 +1,0 @@
-import { createAsyncAction } from "typesafe-actions";
-
-export const pnActivationUpsert = createAsyncAction(
-  "PN_ACTIVATION_UPSERT_REQUEST",
-  "PN_ACTIVATION_UPSERT_SUCCESS",
-  "PN_ACTIVATION_UPSERT_FAILURE"
-)<boolean, boolean, Error>();
